+++ app/src/application/session/IAdminSession.ts
@@ -1,4 +1,4 @@
-import AdminData from '../dto/AdminData';
+import AdminData from '../dto/adminData';
 
 export interface IAdminSession {
   create (adminData: AdminData): string
