+++ app/src/routers/admin/bookRequest.ts
@@ -1,0 +1,2 @@
+/* 予定は未定 */
+export default null;
